chore: move tags to training
This is the only place tags are being used at the moment.
