+++ src/services/core.service.ts
@@ -1,4 +1,4 @@
-import type { City, Country, Tag } from "@/types/core";
+import type { City, Country } from "@/types/core";
 
 import api from "./api";
 
@@ -10,16 +10,6 @@ export default {
 
   fetchCities: async (): Promise<City[]> => {
     const response = await api.get("/api/core/cities");
-    return response.data;
-  },
-
-  fetchTags: async (search: string): Promise<Tag[]> => {
-    const url = "/api/core/tags?";
-    const params = new URLSearchParams();
-
-    params.append("name", search);
-
-    const response = await api.get(url + params.toString());
     return response.data;
   },
 };
