+++ src/types/generic.ts
@@ -1,1 +1,5 @@
 export type StoreStatus = "PENDING" | "FETCHING" | "FETCHED" | "FAILED";
+
+export interface Tag {
+  name: string;
+}
